Upgrade to go-loggregator v3.0.0
Use diego-logging-client as wrapper around go-loggregator to provide an
interface to Diego that fits its needs perfectly while encapsulating the
implementation details of go-loggregator. This design allows for Diego to make
changes to the interface independently of any changes inside go-loggregator.
When go-loggregator is updated, only the diego-logging-client needs to change,
rather than anything that might be using go-loggregator directly.

[#148451433]

Signed-off-by: Brady Love <blove@pivotal.io>